Add files via upload
Changed Interpolation for angle calculation for ground truth at critical point when positive value for gt followed by a negative angle for gt. This requires bringing the negative angle in a positive value only system and then based on the outcome of the interpolation converting it back into a negative angle.